+++ dbt_company_dummy_entity/models/2_prep/prep_product.sql
@@ -1,9 +1,0 @@
-select 
-    id as product_id,
-    name as product_name,
-    category,
-    price,
-    description,
-    created_at,
-    active
-from {{ ref('clean_product') }}