+++ dbt_company_dummy_kimball/models/2_prep/prep_inventory.sql
@@ -1,9 +1,0 @@
-select 
-    id as inventory_id,
-    product_id,
-    current_quantity,
-    reorder_level,
-    reorder_quantity,
-    warehouse_location,
-    last_updated
-from {{ ref('clean_inventory') }}