+++ dbt_company_dummy_kimball/target/compiled/company_dummy_kimball/models/2_prep/prep_department.sql
@@ -1,6 +1,0 @@
-select 
-    id as department_id,
-    name as department_name,
-    description,
-    created_at
-from "company_dummy_kimball"."main"."clean_department"